refactor(web): extract use-scene-projection.ts from SpatialEditor
The post-scene derived memos (keyed, edgePaths, commentChromeBoxes,
selectionMembers, selectionBox, minimapNodes) move verbatim into
their own hook, called immediately after useWorkerScene — every
input (scene, bounds, boxes, canvas, theme, selectedId, extraIds) is
already available there.

buildMinimapNodes moves to minimap.ts in the same commit: the hook
would otherwise import it back from SpatialEditor.tsx, a circular
import. build-minimap-nodes.test.ts's import line updates to match
(the one prescribed mechanical edit this cluster needs).

surfaceKeyed (deps [keyed, draggedCommentId]) and
suppressedBodyNodeIds (deps [editingTextNodeId]) stay in
SpatialEditor.tsx: surfaceKeyed bridges this cluster and the
not-yet-extracted comment-drag state (moving it into either forces a
call-order cycle), and suppressedBodyNodeIds is the scene pipeline's
own input adapter, computed directly above the useWorkerScene call
that consumes it.
